requirements.txt
django
django-redis-sessions==0.6.1
django-extensions
pycrypto
git+https://github.com/VultureProject/djongo
pykeepass==2.8.0